Refactored the codebase to be more organized
